requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
Django==4.2.20
django-appconf==1.1.0
django-apscheduler==0.7.0
django-crontab==0.7.1
django-cryptography==1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.17.0
gunicorn==23.0.0
joblib==1.5.1
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pillow==11.2.1
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
regex==2024.11.6
scikit-learn==1.7.0
scipy==1.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.1
tzlocal==5.3.1
virtualenv==20.29.2
whitenoise==6.9.0
